fix(trusty-audit): the return package refuses when an audited repo's extract database is missing
The owner's ruling on the deliverable (2026-08-18) is the tga collection
database and the rendered reports together. `collect_extract` used to return
`Ok(())` for an audited repository with no matching `extract/<stem>.db` --
both when `extract/` itself did not exist and when it existed but named
nothing for that repository -- so a report could ship with its database
silently missing and nothing in the package said so.

`collect_extract` now refuses the whole assembly with a new, named
`AuditError::MissingExtractDatabase` instead of silently omitting the
database, matching the existing refusal posture for a credential or an
unsafe link. `package.rs` crossed the 500-SLOC production cap adding the new
error's call site, so the credential-scanning apparatus (`CredentialScan`,
`copy_member`, `secret_needles`) moved into a new `package/credential_scan.rs`
submodule -- unchanged in behavior, only in location.

The credential scan already covered every packaged member as an opaque byte
stream regardless of format, so it already scanned the extract database; the
module doc now says so explicitly, alongside what it cannot do (parse the
database's rows, so a secret pasted into a legitimate free-text column still
trips the same refusal a real leak would).

Three integration-test `tga` stubs (chain.rs, cli/registration.rs,
tests/cli_end_to_end.rs) wrote a manifest but never a database file, which
is unrealistic relative to a real `tga audit` and would have failed under
the new refusal; each now also writes `extract/<stem>.db`, deriving the path
from `--output` since `out/` and `extract/` are always siblings under the
work-dir root.

Closes #5862
